api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
qdrant-client==1.8.0
numpy==1.26.0
httpx==0.27.0
openai==1.35.0
rank-bm25==0.2.2
python-multipart==0.0.9
pytest==8.1.1
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-propagator-b3==1.23.0
opentelemetry-propagator-jaeger==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-requests==0.44b0
opentelemetry-instrumentation-logging==0.44b0
langfuse==2.36.0
sentence-transformers==2.7.0
psutil==5.9.8
redis==5.0.1
prometheus-client==0.20.0
PyMuPDF==1.23.0
python-magic==0.4.27
cachetools==5.3.0